matrixMul works with local assemblies.
Lots of modifications for this commit.
- All CUDA Runtime functions within matrixMul added to both the interposing
	library and localsink, including some __cudaRegister*() functions.
- Provided a workaround to enabling a multi-process backend. See function
	hack_getCudaInformation in assembly.c
- Added some documentation for limitations/assumptions in README
- Added feature/implementation list for some items in TODO
- Added CUDA Runtime and Driver versions to participant structures
- Added complex state/functions needed to support setDevice/getDevice
	See functions get_thread_association and set_thread_association in
	assembly.c
- Fixed a bug in compose_assembly to use 2 vgpus
- Added some assumption notes to assembly_rpc()
- Moved execution of CUDA packets to local vgpus into nv_exec_pkt()
	(I think this should be converted to a jump table instead)
- Removed some useless comments in packetheader.h
- Modified some functions in libciutils.c to be called from libci.c in order to
	support interposition of __cudaRegister*() state into shared memory

Still TODO
- libregistration does not clean up properly
- Supports only one __cudaRegisterFatBinary invocation

+++ include/packetheader.h
@@ -161,16 +161,16 @@ typedef struct {
 // Arguments will be filled in the order as in function declaration
 // Currently this union has the possible arguments seen in common CUDA calls
 typedef union args {
-	int arr_argii[4];                   // int[4]
-	unsigned int arr_arguii[4];         // unsigned int[4]
-	int64_t argll;                      // signed long long (64-bit)
-	uint64_t argull;                    // unsigned long long (64-bit)
-	float argf;                         // float
-	void *argp;                         // void pointer
-	void **argdp;                       // double void pointer
-	char *argcp;                        // char pointer
-	size_t arr_argi[2];                 // size_t[2]
-	uint64_t arr_argui[2];              // unsigned int64[2]
+	int arr_argii[4];
+	unsigned int arr_arguii[4];
+	long long argll;
+	unsigned long long argull;          // for pointers and such
+	float argf;
+	void *argp;
+	void **argdp;
+	char *argcp;
+	size_t arr_argi[2];
+	unsigned long long arr_argui[2];
 	tf_args_t tf_args;                  // ?
 	dim3 arg_dim;                       // 3D point
 	cudaStream_t arg_str;
